==> code/00_run_order.R <==
source("01_requirements.R")

source("01_check_used_functions.R")

#GET
source("02_subregions_glottocodes_unnesting.R")

source("02_get_glottolog_language_table.R")

source("02_ecoClimate_extract.R")

source("02_get_modis.R")

#WRANGLE
source("03a_combine_data.R")

source("03b_scale_data.R")

source("03c_environ_PCA.R")

source("03d_country_pop_vs_lg.R")

source("03e_spatial_dists.R")

source("03f_phylo_dists.R")

source("03g_NANO_contact_map.R")

#PLOT
source("04_data_maps.R")

source("04_map_polygon_groups.R")

source("04_variable_SPLOM.R")

#ANALYZE

source("05a_analysis_brms.R")

source("05b_model_fit_comparison.R")

source("05c_best_models.R")

source("06_prep_appendices.R")
